requirements.txt
psycopg2-binary==2.9.9
openai==1.55.3
PyMuPDF==1.23.7
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.1
httpx==0.27.2